Migrating shadows to a unrasteriser architecture
Former-commit-id: b8471243a52e583b5d259b5fcba4a5d847c294e2 [formerly a359b75d27a86b531491fc3cb18f3292435f0c39]
Former-commit-id: baf4e36722ae488d7ce7a0546620d8251651e29d
Former-commit-id: 69790e30e53246d3854414a0624b27a5fe0985ce
Former-commit-id: ead6f5ba1b4d19b3df8bdedfd4704b5da90c46ef

+++ src/engine/planar/textureprojector.cpp
@@ -15,11 +15,11 @@ void TextureProjector::generate_uv_projector(const Triangle& projected_triangle,
 
   // Projected triangle vertices as base of the screen (origin)
   Basis2 screen_basis ({
-               {double(projected_triangle.b.X - projected_triangle.a.X),
-                  double(projected_triangle.b.Y - projected_triangle.a.Y)},
+               {projected_triangle.b.X - projected_triangle.a.X,
+                projected_triangle.b.Y - projected_triangle.a.Y},
 
-               {double(projected_triangle.c.X - projected_triangle.a.X),
-                  double(projected_triangle.c.Y - projected_triangle.a.Y)},
+               {projected_triangle.c.X - projected_triangle.a.X,
+                projected_triangle.c.Y - projected_triangle.a.Y},
              });
 
   // Generate matrix to change between basis
